Noetic Docker/rl_PaP/requirements.txt
actionlib==1.14.0
addict==2.4.0
aiofiles==22.1.0
aiosqlite==0.20.0
angles==1.9.13
anyio==4.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==3.0.0
attrs==25.3.0
autobahn==17.10.1
Automat==0.8.0
babel==2.17.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
bleach==6.1.0
blinker==1.4
bondpy==1.8.6
breezy==3.0.2
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==1.0.0
catkin-pkg-modules==1.0.0
catkin-tools==0.9.4
cbor==1.0.0
certifi==2019.11.28
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
Click==7.0
colorama==0.4.3
comm==0.2.2
configobj==5.0.6
constantly==15.1.0
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cryptography==2.8
cupshelpers==1.0
cv-bridge==1.16.2
cycler==0.10.0
Cython==0.29.14
dbus-python==1.2.16
debugpy==1.8.14
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
Deprecated==1.2.7
deprecation==2.1.0
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distro==1.4.0
docutils==0.16
dulwich==0.19.15
dynamic-reconfigure==1.7.3
dynamixel-sdk==3.8.0
empy==3.3.2
entrypoints==0.3
exceptiongroup==1.2.2
executing==2.2.0
fastimport==0.9.8
fastjsonschema==2.21.1
gazebo-plugins==2.9.3
gazebo-ros==2.9.3
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
gpg==1.13.1
html5lib==1.0.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
image-geometry==1.16.2
importlib-metadata==8.5.0
importlib-resources==6.4.5
incremental==16.10.1
interactive-markers==1.12.0
ipykernel==6.29.5
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==8.1.6
jedi==0.19.2
jinja2==3.1.6
joblib==1.4.2
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
json5==0.12.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-client==8.6.3
jupyter-core==5.7.2
jupyter-events==0.10.0
jupyter-packaging==0.12.3
jupyter-server==2.14.2
jupyter-server-fileid==0.9.3
jupyter-server-terminals==0.5.3
jupyter-server-ydoc==0.8.0
jupyter-ydoc==0.2.5
jupyterlab==3.6.8
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.14
#kdl-parser-py==1.14.2
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
laser-geometry==1.6.7
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
llvmlite==0.41.1
lxml==4.5.0
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
MarkupSafe==2.1.5
matplotlib==3.1.2
matplotlib-inline==0.1.7
message-filters==1.17.0
mistune==3.1.3
more-itertools==4.2.0
moveit-commander==1.1.16
moveit-core==1.1.16
moveit-ros-planning-interface==1.1.16
moveit-ros-visualization==1.1.16
movement-primitives==0.9.0
mpi4py==3.0.3
nbclassic==1.2.0
nbclient==0.10.1
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.10.4
nose==1.3.7
notebook==6.5.7
notebook-shim==0.2.4
numba==0.58.1
numpy==1.24.4
oauthlib==3.1.0
olefile==0.46
open3d==0.13.0
osrf-pycommon==2.0.2
overrides==7.7.0
packaging==24.2
pandas==2.0.3
pandocfilters==1.5.1
paramiko==2.6.0
parso==0.8.4
pbr==6.1.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil-resolve-name==1.3.10
platformdirs==4.3.6
prometheus-client==0.21.1
prompt-toolkit==3.0.51
protobuf==3.6.1
psutil==5.5.1
ptyprocess==0.7.0
pure-eval==0.2.3
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pyassimp==4.1.4
pycairo==1.16.2
pycparser==2.22
pycryptodomex==3.6.1
pycups==1.9.73
pydot==1.4.1
PyGithub==1.43.7
pygments==2.19.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.14.1
pyRFC3339==1.1
pyserial==3.5
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-gitlab==2.0.1
python-gnupg==0.4.5
python-json-logger==3.3.0
python-qt-binding==0.4.4
python-snappy==0.5.3
pytransform3d==3.14.0
PyTrie==0.2
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
referencing==0.35.1
requests==2.32.3
requests-unixsocket==0.2.0
resource-retriever==1.12.8
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roman==2.0.0
rosbag==1.17.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosdep==0.25.1
rosdep-modules==0.25.1
rosdistro==1.0.1
rosdistro-modules==1.0.1
rosgraph==1.17.0
rosinstall==0.7.8
rosinstall-generator==0.1.23
roslaunch==1.17.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.17.0
rosmake==1.15.8
rosmaster==1.17.0
rosmsg==1.17.0
rosnode==1.17.0
rosparam==1.17.0
rospkg==1.6.0
rospkg-modules==1.6.0
rospy==1.17.0
rosservice==1.17.0
rostest==1.17.0
rostopic==1.17.0
rosunit==1.15.8
roswtf==1.17.0
rpds-py==0.20.1
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.12
rqt-dep==0.4.12
rqt-graph==0.4.14
rqt-gui==0.5.3
rqt-gui-py==0.5.3
rqt-image-view==0.4.17
rqt-launch==0.4.9
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-msg==0.4.10
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.11
rqt-publisher==0.4.10
rqt-py-common==0.5.3
rqt-py-console==0.4.10
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-robot-steering==0.5.12
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.0
rqt-service-caller==0.4.10
rqt-shell==0.4.11
rqt-srv==0.4.9
rqt-tf-tree==0.6.4
rqt-top==0.4.10
rqt-topic==0.4.13
rqt-web==0.4.10
rviz==1.14.25
scikit-learn==1.3.2
scipy==1.10.1
SecretStorage==2.3.1
Send2Trash==1.8.3
sensor-msgs==1.13.1
service-identity==18.1.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
smach==2.5.3
smach-ros==2.5.3
smclib==1.8.6
sniffio==1.3.1
soupsieve==1.9.5
srdfdom==0.6.4
stack-data==0.6.3
systemd-python==234
terminado==0.18.1
testresources==2.0.1
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
threadpoolctl==3.5.0
tinycss2==1.2.1
tomli==2.2.1
tomlkit==0.13.2
topic-tools==1.17.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
Twisted==18.9.0
txaio==2.10.0
typing-extensions==4.13.2
tzdata==2025.2
u-msgpack-python==2.1
urdfdom-py==0.4.6
urllib3==1.25.8
vcstools==0.1.42
wadllib==1.3.3
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.14
wrapt==1.11.2
wsaccel==0.6.2
wstool==0.1.18
xacro==1.14.20
y-py==0.6.2
ypy-websocket==0.8.4
zipp==3.20.2
zope.interface==4.7.1